+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 - Added thread and message focus mode - [#129](https://github.com/hkdb/aerion/issues/129)
 - Added contact circle enable/disable
 - Added French translation
-- Added keyboard expand/collapse of focused messages
+- Fixed keyboard expand/collapse of focused messages
 - Fixed untranslated drop down menu items
 - Fixed iCloud Empty Trash - [#136](https://github.com/hkdb/aerion/issues/136)
 - Added ESLint for frontend linting